# service/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : { },
  "tags" : [ {
    "name" : "auth"
  }, {
    "name" : "exercise"
  }, {
    "name" : "personal-website"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/auth/verify" : {
      "get" : {
        "tags" : [ "auth" ],
        "summary" : "Verify a JWT token",
        "description" : "",
        "operationId" : "verifyJWT",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/auth/sign" : {
      "get" : {
        "tags" : [ "auth" ],
        "summary" : "Sign a new JWT token with username and password information",
        "description" : "",
        "operationId" : "getJWT",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/exercise/exercise-summary" : {
      "get" : {
        "tags" : [ "exercise" ],
        "summary" : "Getting exercise summary",
        "description" : "",
        "operationId" : "getExerciseSummary",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ExerciseSummaryDTO"
            }
          }
        }
      }
    },
    "/exercise/gym-session" : {
      "post" : {
        "tags" : [ "exercise" ],
        "summary" : "Add gym session",
        "description" : "",
        "operationId" : "addGymSession",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "runningDistanceInMiles",
          "in" : "query",
          "required" : false,
          "type" : "number",
          "format" : "float"
        }, {
          "name" : "durationInMinutes",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/exercise/strava" : {
      "get" : {
        "tags" : [ "exercise" ],
        "summary" : "Get authorization code for Strava auth",
        "description" : "",
        "operationId" : "getStravaCode",
        "parameters" : [ {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "exercise" ],
        "summary" : "add strava code",
        "description" : "",
        "operationId" : "addStravaCode",
        "parameters" : [ {
          "name" : "code",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/exercise/strava-status" : {
      "get" : {
        "tags" : [ "exercise" ],
        "summary" : "getStravaStatus",
        "description" : "",
        "operationId" : "getStravaStatus",
        "parameters" : [ {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "exercise" ],
        "summary" : "setStravaStatus",
        "description" : "",
        "operationId" : "setStravaStatus",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/personal-website/base-website" : {
      "put" : {
        "tags" : [ "personal-website" ],
        "summary" : "Update base website info",
        "description" : "",
        "operationId" : "updateBaseWebsiteInfo",
        "parameters" : [ {
          "name" : "firstParagraphOfAboutMe",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "secondParagraphOfAboutMe",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "listOfTechnicalLanguages",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "listOfFrameworks",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "listOfTools",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "listOfLanguages",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "resumePath",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/personal-website/projects" : {
      "post" : {
        "tags" : [ "personal-website" ],
        "summary" : "Add a new project and get its id",
        "description" : "",
        "operationId" : "addProjectInfo",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "description",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "funFact",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "url",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "imagePath",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "year",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "personal-website" ],
        "summary" : "Get website info using the model WebsiteInfo",
        "description" : "",
        "operationId" : "updateProjectInfo",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "description",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "funFact",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "url",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "imagePath",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "year",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "bearerAuth",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/personal-website/website" : {
      "get" : {
        "tags" : [ "personal-website" ],
        "summary" : "Get full website info using the model FullWebsiteInfo",
        "description" : "",
        "operationId" : "getWebsiteInfo",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FullWebsiteInfo"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "ExerciseSummaryDTO" : {
      "type" : "object",
      "properties" : {
        "lengthOfStreakInDays" : {
          "type" : "integer",
          "format" : "int32"
        },
        "milesRanToday" : {
          "type" : "number",
          "format" : "float"
        },
        "numberOfMinutesInGym" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "FullWebsiteInfo" : {
      "type" : "object",
      "required" : [ "firstParagraphOfAboutMe", "id", "listOfFrameworks", "listOfLanguages", "listOfTechnicalLanguages", "listOfTools", "projectInfoList", "resumePath", "secondParagraphOfAboutMe" ],
      "properties" : {
        "firstParagraphOfAboutMe" : {
          "type" : "string"
        },
        "secondParagraphOfAboutMe" : {
          "type" : "string"
        },
        "listOfTechnicalLanguages" : {
          "type" : "string"
        },
        "listOfFrameworks" : {
          "type" : "string"
        },
        "listOfTools" : {
          "type" : "string"
        },
        "listOfLanguages" : {
          "type" : "string"
        },
        "resumePath" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "projectInfoList" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/ProjectInfo"
          }
        }
      }
    },
    "ProjectInfo" : {
      "type" : "object",
      "required" : [ "description", "funFact", "id", "imagePath", "name", "type", "url", "year" ],
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "funFact" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "imagePath" : {
          "type" : "string"
        },
        "year" : {
          "type" : "integer",
          "format" : "int32"
        },
        "id" : {
          "type" : "string"
        }
      }
    }
  }
}